runtime: move RuntimeProxy code to submodule
No code changes, just moving code around and fixing `use`
statements, plus adding one `pub(super)` for `RuntimeProxy::runtime`.

For #1022.
